Add the argument "-static" for Static Library.
If your library folder have both of the Static and the Dynamic libraries, the gcc will link Dynamic Library by default.
You need to add the keyword "-static" if you want to link Static Library.

But the keyword "-shared" ISN'T to mean "link to Dynamic Library"; its meaning is to "create a Dynamic Library" conversely.